Refactor grib tests (working *without* iris_grib).
