+++ conda-requirements.txt
@@ -20,7 +20,7 @@ nose
 pep8=1.5.7
 sphinx
 iris_sample_data
-iris_grib
+#iris_grib
 
 # Optional iris dependencies
 ecmwf_grib
